sp1_bls12_381::fp

Struct Fp

pub struct Fp(/* private fields */);

Implementations

impl Fp

pub const fn zero() -> Fp

Returns zero, the additive identity.

pub const fn one() -> Fp

Returns one, the multiplicative identity.

pub fn is_zero(&self) -> Choice

pub fn from_bytes(bytes: &[u8; 48]) -> CtOption<Fp>

Attempts to convert a big-endian byte representation of a scalar into an Fp, failing if the input is not canonical.

pub fn to_bytes(&self) -> [u8; 48]

Converts an element of Fp into a byte representation in big-endian byte order.

pub fn random(rng: impl RngCore) -> Fp

pub fn lexicographically_largest(&self) -> Choice

Returns whether or not this element is strictly lexicographically larger than its negation.

pub const fn from_raw_unchecked(v: [u64; 6]) -> Fp

Constructs an element of Fp without checking that it is canonical.

pub fn pow_vartime(&self, by: &[u64; 6]) -> Self

Although this is labeled “vartime”, it is only variable time with respect to the exponent. It is also not exposed in the public API.

pub fn sqrt(&self) -> CtOption<Self>

pub fn invert(&self) -> CtOption<Self>

pub fn add(&self, rhs: &Fp) -> Fp

pub fn neg(&self) -> Fp

pub fn sub(&self, rhs: &Fp) -> Fp

pub fn mul(&self, rhs: &Fp) -> Fp

pub fn square(&self) -> Self

Squares this element.