sp1_bls12_381

Module g1

This module provides an implementation of the $\mathbb{G}_1$ group of BLS12-381.

Structs

G1Affine

This is an element of $\mathbb{G}_1$ represented in the affine coordinate space. It is ideal to keep elements in this representation to reduce memory usage and improve performance through the use of mixed curve model arithmetic.

G1Compressed

G1Projective

This is an element of $\mathbb{G}_1$ represented in the projective coordinate space.

G1Uncompressed

Constants

BETA

A nontrivial third root of unity in Fp

Functions

endomorphism